refactor(settings): extract WeightBackupCodec from SerializationService
