+++ lib/features/settings/data/serialization_service.dart
@@ -1,9 +1,7 @@
 import 'dart:io';
-import 'package:archive/archive.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/widgets.dart';
-import 'package:food_locker/core/csv_serializer.dart';
-import 'package:food_locker/features/weight/data/weight.dart';
+import 'package:food_locker/features/settings/data/weight_backup_codec.dart';
 import 'package:food_locker/features/weight/data/weight_repository.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:provider/provider.dart';
@@ -28,11 +26,11 @@ class SerializationService {
   Future<void> exportData(BuildContext context) async {
     final weightRepo = context.read<WeightRepository>();
 
-    final zipData = createExportArchive(
+    final zipData = const WeightBackupCodec().encode(
       weightRepo.getAllWeights(),
     );
 
-    if (zipData == null) return;
+    if (zipData.isEmpty) return;
 
     final tempDir = await getTemporaryDirectory();
     final zipFile = File('${tempDir.path}/${generateZipFileName()}');
@@ -56,77 +54,10 @@ class SerializationService {
     final file = File(filePath);
     final bytes = await file.readAsBytes();
 
-    await importFromArchive(bytes, weightRepo);
-  }
-
-  @visibleForTesting
-  List<int>? createExportArchive(List<Weight> weights) {
-    final weightCsv = generateWeightCsv(weights);
-
-    final archive = Archive()
-      ..addFile(
-        ArchiveFile(_weightFileName, weightCsv.length, weightCsv.codeUnits),
-      );
-
-    return ZipEncoder().encode(archive);
-  }
-
-  @visibleForTesting
-  Future<void> importFromArchive(
-    List<int> zipBytes,
-    WeightRepository weightRepo,
-  ) async {
-    final archive = ZipDecoder().decodeBytes(zipBytes);
-
-    // Clear existing data
+    final weights = const WeightBackupCodec().decode(bytes);
     await weightRepo.clear();
-
-    for (final file in archive) {
-      if (file.isFile) {
-        final content = String.fromCharCodes(file.content as List<int>);
-        if (file.name == _weightFileName) {
-          importWeightFromCsv(content, weightRepo);
-        }
-      }
-    }
-  }
-
-  @visibleForTesting
-  String generateWeightCsv(List<Weight> weights) {
-    final items = weights
-        .map((w) => {
-              'date': w.date.toIso8601String(),
-              'value': w.value,
-              'unit': w.unit.name,
-            })
-        .toList();
-    return CsvSerializer.toCSV(items);
-  }
-
-  @visibleForTesting
-  void importWeightFromCsv(String csv, WeightRepository repo) {
-    final items = CsvSerializer.fromCSV(csv);
-    for (final item in items) {
-      final dateStr = item['date'] as String?;
-      final valueStr = item['value']?.toString();
-      final unitStr = item['unit'] as String?;
-
-      if (dateStr == null || valueStr == null) continue;
-
-      final date = DateTime.tryParse(dateStr);
-      final value = double.tryParse(valueStr);
-
-      if (date != null && value != null) {
-        WeightUnit unit = WeightUnit.kilograms;
-        if (unitStr != null) {
-          try {
-            unit = WeightUnit.values.byName(unitStr);
-          } catch (e) {
-            debugPrint('Error parsing weight unit: $unitStr, $e');
-          }
-        }
-        repo.saveWeight(Weight(date: date, value: value, unit: unit));
-      }
+    for (final weight in weights) {
+      await weightRepo.saveWeight(weight);
     }
   }
 }
